refactor(staking): rename feature components to PascalCase
