+++ packages/staking/.eslintignore
@@ -1,1 +1,1 @@
-src/features/browse-pools/data.json
+src/features/BrowsePools/data.json
