+++ packages/staking/src/features/BrowsePools/PortfolioBar.tsx
