+++ packages/staking/src/features/BrowsePools/StakePoolsTable/StakePoolItemBrowser/StakePoolItemBrowser.module.scss
