+++ packages/staking/src/features/BrowsePools/StakePoolsTable/StakePoolsTable.module.scss
