+++ packages/staking/src/features/Drawer/preferences/StepPreferencesContent.tsx
@@ -2,7 +2,7 @@
 import { Wallet } from '@lace/cardano';
 import { Box, ControlButton, Flex, PIE_CHART_DEFAULT_COLOR_SET, PieChartColor, Text } from '@lace/ui';
 import { useTranslation } from 'react-i18next';
-import { DelegationCard, DelegationStatus } from '../../delegation-card';
+import { DelegationCard, DelegationStatus } from '../../DelegationCard';
 import { useOutsideHandles } from '../../outside-handles-provider';
 import {
   DelegationFlow,
