+++ packages/staking/src/features/staking/StakingView.tsx
@@ -1,7 +1,7 @@
 import { Box, Text } from '@lace/ui';
 import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
-import { BrowsePools } from '../browse-pools';
+import { BrowsePools } from '../BrowsePools';
 import { Drawer } from '../drawer';
 import { ChangingPreferencesModal, MultidelegationBetaModal } from '../modals';
 import { useOutsideHandles } from '../outside-handles-provider';
